% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJ_09_variation.R
\name{AJ_09_variation}
\alias{AJ_09_variation}
\title{Ait-Sahalia & Jacod Variation (2009)}
\usage{
AJ_09_variation(DATA, k = 3, p = 4, gamma = 2, a = 10)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{k}{The degree of subsampling the frequency in the multipower variation. Defaults to 3.}

\item{p}{The power for computing the multipower variation. Defaults to 4.}

\item{gamma}{Value >1 for amplifying the truncation. Defaults to 2.}

\item{a}{The level of truncation expressed in terms of numbers of stdev of the continuous part.  Defaults to 10.}
}
\value{
Returns the variation estimation
}
\description{
Calculate sigma_hat according to Ait-Sahalia & Jacod (2009) "Estimating the degree of activity of jumps in high frequency data"
}
