% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJ_JumpTest_2012.R
\name{AJ_JumpTest}
\alias{AJ_JumpTest}
\title{Ait-Sahalia & Jacod Jump Test}
\usage{
AJ_JumpTest(DATA, alpha = 0.1)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}
}
\value{
Returns the parameter grid and limits for the tests for jumps and for finite jump activity.
}
\description{
Calculates the jump test statistic and test for finite activity as per Ait-Sahalia & Jacod (2012)
}
