% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJL_Jump_Test_2012.R
\name{AJL_JumpTest_2012}
\alias{AJL_JumpTest_2012}
\title{Ait-Sahalia, Jacod & Li Jump Test}
\usage{
AJL_JumpTest_2012(
  DATA,
  deltan = 1/365.25/86400,
  kn = 100,
  T_large = 1/365.25,
  trunc_level = 5,
  variance_flag = 1,
  sign_level = 0.05
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{deltan}{The frequency of the data. Defaults to 1/365.25/86400 (i.e. 1 second in a 365.25 trading days interval).}

\item{sign_level}{}

\item{kn:}{The block size for preaveraging. Defaults to 100.}

\item{T_large:}{The length of T. Defaults to 1/365.25}
}
\value{
DT_vvec:

S_stdc:

sigmabar

un

variancehat

Mstargg

Mstargh

Mstarhh

sign_level

pvalue

cv

Jump_indicator
}
\description{
Calculates the jump test statistic as per Ait-Sahalia, Jacod & Li (2012)
}
