% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Data Cleaning}
\usage{
clean_data(DATA, id = NA)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{id}{Optional argument in case Blockchain Research Center data as in the example is used. Defaults to NA.}
}
\value{
Returns the cleaned data
}
\description{
Helper function for preprocessing the data for calculating jump tests
}
