% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_crypto_data.R
\name{download_crypto_data}
\alias{download_crypto_data}
\title{Download crypto data}
\usage{
download_crypto_data(LINK, PATH = "./data/raw/csv_dump/")
}
\arguments{
\item{LINK}{a link from the BRC.}

\item{PATH}{the path where the data should be downloaded to.}
}
\description{
Helper function to download data from the BRC
}
