% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJL_Jump_Test_2012_functions.R
\name{gfun}
\alias{gfun}
\title{Ait-Sahalia, Jacod & Li Jump Test Functions}
\usage{
gfun(x)
}
\arguments{
\item{mfun}{m(r) = E|N(0,1)|^r, see Jacod and Ait-Sahalia 2006 equation (10)}

\item{Vvec}{The jth element, j=0,...,p/2, is V(Y,g,p-2j,j) = sum |Ybar|^(p-2j) Yhat^j}

\item{rho_tilde}{The rho_tilde matrix in AJL2009}

\item{...}{}
}
\description{
Helper functions for computing the Ait-Sahalia, Jacod & Li (2012) test
}
