% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacod_preaveraging.R
\name{jacod_preaveraging}
\alias{jacod_preaveraging}
\title{Jacod Preaveraging (2009, 2010)}
\usage{
jacod_preaveraging(DATA, kn = 100)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{kn}{The degree of subsampling during preaveraging. Defaults to 100.}
}
\value{
Returns the variation estimation
}
\description{
Calculate sigma_hat according to the pre-averaging approach Jacod et al. (2009, 2010)
}
