% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_JumpTest_2012.R
\name{LM_JumpTest}
\alias{LM_JumpTest}
\title{Lee & Mykland Jump Test}
\usage{
LM_JumpTest(DATA, variation_estimate = "Jacod_10")
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{variation_estimate}{How should sigma_hat be estimated? Options are "HP_13" for Hautsch & Podolskij (2013) "Modulated realized covariance"; "Jacod_10" for Jacod et al. (2010) "Pre-averaging"; "AJ_09" for Ait-Sahalia & Jacod (2009) "Jump Activity". Defaults to "Jacod_10".}
}
\value{
Returns a list with both the ICC and the ISVM result.

t: The timestamps on `G_n_kM``

date: The corresponding date

id: in case the info is available, on which exchange was the data point observed?

s: the ticker symbol

count: the count of the observation in the full dataset

P_hat_tj: `P_hat_tj` calculated as in the original paper. Averaged subsampled observations over blocks.

exp_P_hat_tj:  `e^(P_hat_tj)` to approximate the current price.

exp_P_hat_tj_minus1 Lagged value of `e^(P_hat_tj)`. This shows the price of the previous observation.

k: calculated as in the original paper. Indicates the degree of contamination with market microstructure noise in the data.

M: calculated as in the original paper. The block size for subsampling.

kM: `k*M`. Shows how many observations are in each block.

C: calculated as in the original paper.

qhat: calculated as in the original paper.

sigmahat: calculated as in the original paper.

Vn: The daily variation estimated from the data.

plim_Vn: The asymptotic daily variation.

A_n: calculated as in the original paper.

B_n calculated as in the original paper.

L_t_j: calculated as in the original paper. The returns between the averages of subsampled observations over blocks.

Chi_t_j: calculated as in the original paper. For calculating the test statistic.

sign_level: Significance level for the test statistic.

betastar: Value for obtaining the critical value.

criticalvalue: The critical value for rejecting the null.

Jump_indicator: Either -1,0, or 1. Where 0 = no jump and the sign indicates the direction of the jump.
}
\description{
Calculates the jump test statistic as per Lee & Mykland (2012)
}
