% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_split.R
\name{load_and_split}
\alias{load_and_split}
\title{Load and split}
\usage{
load_and_split(crypto_data_path = NA, clean_data_path = "./data/clean/")
}
\description{
Helper function to process data from the BRC
}
