% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_data_directory.R
\name{make_clean_data_directory}
\alias{make_clean_data_directory}
\title{Make clean data directory}
\usage{
make_clean_data_directory(PATH = "./data/clean/")
}
\description{
Helper function to create data directories
}
