% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regular_return_file.R
\name{make_regular_return_file}
\alias{make_regular_return_file}
\title{Make regular return file}
\usage{
make_regular_return_file(
  DATA,
  delts = c(1, 5, 10, 15),
  full_day = 0.95 * 86400,
  IMPUTATION = FALSE,
  T_large = 1,
  ALIGN = TRUE,
  FREQ = 1
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{delts}{Frequencies that should be tested for subsampling the regularized data, where e.g. a value of 5 means that the data is subsampled every 5 seconds. Defaults to \code{c(5, 10, 15)}.}

\item{full_day}{How many observations in seconds does a full day have? Defaults to \code{0.95*86400}.}

\item{IMPUTATION}{Should missing values be imputed? If TRUE, then \code{na.locf} will be used. Defaults to FALSE.}

\item{T_large}{The number of trading days. Defaults to 1 for one observed day.}

\item{ALIGN}{Should the data be aligned to round timestamps? Defaults to TRUE.}

\item{FREQ}{The highest possible frequency. Defaults to 1 for a frequency of 1 second.}
}
\value{
Outputs a regularized return file with appropriate frequency.
}
\description{
Helper function to create a regular return file
}
