% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_return_file.R
\name{make_return_file}
\alias{make_return_file}
\title{Make return file}
\usage{
make_return_file(DATA, FREQ = 1, IMPUTATION = FALSE, CUTOFF = 10, ALIGN = TRUE)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{FREQ}{The highest possible frequency. Defaults to 1 for a frequency of 1 second.}

\item{IMPUTATION}{Should missing values be imputed? If TRUE, then \code{na.locf} will be used. Defaults to FALSE.}

\item{CUTOFF}{The number of standard deviations of log returns around which data should be kept. Defaults to 10, s.t. observations outside +-10 SD will be omitted.}

\item{ALIGN}{Should the data be aligned to round timestamps? Defaults to TRUE.}
}
\value{
Outputs a cleaned return file.
}
\description{
Helper function to create a regular return file
}
