% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files.R
\name{preprocess_jump_data}
\alias{preprocess_jump_data}
\title{Preprocess LM jump test result}
\usage{
preprocess_jump_data(
  DATA,
  subs_date = NA,
  sign_level = 0.01,
  split_waves = FALSE,
  wave_dist = NA,
  wave_size = 5,
  bonferroni = TRUE,
  DIFF_ID = FALSE,
  asymptotic_variation = FALSE,
  hourly = FALSE
)
}
\arguments{
\item{DATA}{The output from \code{LM_JumpTest_2012}.}

\item{subs_date}{A vector with dates can be provided to subset from. Defaults to NA.}

\item{sign_level}{The significance level for detecting jumps. Defaults to 0.01.}

\item{split_waves}{Should the data be split into jump waves? Defaults to FALSE.}

\item{wave_dist}{If \code{split_waves = TRUE}, where to set the splits? E.g. if no jump for longer than \code{wave_dist = 5}, the wave is over. This is adaptive w.r.t. to minutely / secondly frequency. Defaults to NA.}

\item{wave_size}{Lower threshold of jumps needed to be in a wave for it to be considered? Defaults to 5, i.e. only waves with at least 5 jumps will be kept in the end. Set to 1 for all jumps to be included.}

\item{bonferroni}{Should a Bonferroni correction be performed? Defaults to TRUE.}

\item{DIFF_ID}{Should the data be evaluated by exchange and symbol or just by exchange? Defaults to FALSE, i.e. jumps are calculated per symbol.}

\item{asymptotic_variation}{Should the asymptotic variation be used for calculating the test statistic? Defaults to FALSE, s.t. the empirical variation is used.}

\item{hourly}{Is the test statistic calculated hourly? This can be useful for very high frequency data. Defaults to FALSE.}
}
\description{
Convenience functionalities for subsetting the jump results, changing the significance level, splitting the jumps into different waves, adding a Bonferroni correction, splitting the data by exchange / symbol, and modifying the variation measure for the test statistic
}
