% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bounceback.R
\name{remove_bounceback}
\alias{remove_bounceback}
\title{Remove bounceback outliers}
\usage{
remove_bounceback(DATA, BOUNCE_CUTOFF = 0.001, IMPUTATION = FALSE)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{BOUNCE_CUTOFF}{The threshold of log return size for which bounceback outliers should be removed. Defaults to 0.001, s.t. smaller returns are not checked for bounceback outliers.}

\item{IMPUTATION}{Should missing values be imputed? If TRUE, then \code{na.locf} will be used. Defaults to FALSE.}
}
\value{
Returns the data without bounceback outliers.
}
\description{
Helper function to remove bounceback outliers.
}
