% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_consecutive_jumps.R
\name{remove_consecutive_jumps}
\alias{remove_consecutive_jumps}
\title{Remove consecutive jumps}
\usage{
remove_consecutive_jumps(
  DATA,
  RANGE_OBS = 3,
  split_vars = c("date", "hour", "s")
)
}
\arguments{
\item{DATA}{The output from \code{LM_JumpTest_2012}.}

\item{RANGE_OBS}{How many observations away must a jump be to be considered non-consecutive? Defaults to 3, s.t. all jumps in the next 3 blocks will be removed.}

\item{split_vars}{By which variables should the data be split? Defaults to \code{c("date", "hour", "s")}.}
}
\value{
Returns the data without consecutive jumps
}
\description{
Helper function to remove consecutive jumps
}
