% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_id.R
\name{split_by_id}
\alias{split_by_id}
\title{Split by ID}
\usage{
split_by_id(
  DATA,
  IMPUTATION = FALSE,
  T_large = 1,
  full_day = 0.95 * 86400,
  delts = c(5, 10, 15),
  FREQ = 1,
  ALIGN = TRUE
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{IMPUTATION}{Should missing values be imputed? If TRUE, then \code{na.locf} will be used. Defaults to FALSE.}

\item{T_large}{The number of trading days. Defaults to 1 for one observed day.}

\item{full_day}{How many observations in seconds does a full day have? Defaults to \code{0.95*86400}.}

\item{delts}{Frequencies that should be tested for subsampling the regularized data, where e.g. a value of 5 means that the data is subsampled every 5 seconds. Defaults to \code{c(5, 10, 15)}.}

\item{ALIGN}{Should the data be aligned to round timestamps? Defaults to TRUE.}
}
\description{
Helper function to split data by ID and make it regular for computing the test statistic.
}
