% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJL_Jump_Test_2012.R
\name{AJL_JumpTest_2012}
\alias{AJL_JumpTest_2012}
\title{Ait-Sahalia, Jacod & Li Jump Test}
\usage{
AJL_JumpTest_2012(
  DATA,
  deltan = 1/365.25/86400,
  kn = 100,
  T_large = 1/365.25,
  trunc_level = 5,
  variance_flag = TRUE,
  sign_level = 0.05
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{deltan}{The frequency of the data. Defaults to 1/365.25/86400 (i.e. 1 second in a 365.25 trading days interval).}

\item{kn}{The block size for preaveraging. Defaults to 100.}

\item{T_large}{The length of T. Defaults to 1/365.25}

\item{variance_flag}{Should the variance be estimated? Defaults to TRUE.}

\item{sign_level}{Significance level for the test statistic. Defaults to 0.05.}

\item{trunc_level:}{Value for the truncation levels for estimating asymptotic variances. Defaults to 5.}
}
\value{
DT_vvec: The V vectors from pre-averaging

S_stdc: The standardized noise robust test statistic

sigmabar: The average volatility

un: The truncation level

variancehat: The estimated variance

Mstargg: The Mstar values as in the original paper

Mstargh: The Mstar values as in the original paper

Mstarhh: The Mstar values as in the original paper

sign_level: The signficance level

pvalue: The p-value

cv: The critical value

Jump_indicator: The jump indicator
}
\description{
Calculates the jump test statistic as per Ait-Sahalia, Jacod & Li (2012)
}
